chore(js): convert JS to AMD for several plugins
BREAKING CHANGE:

If you were loading any core plugins with, for example, `elgg_load_js('elgg.discussion')`,
you will need to change that to use AMD, for example, `elgg_require_js('elgg/discussion')`.
