+++ mod/ckeditor/views/default/js/elgg/ckeditor/init.js
@@ -1,0 +1,8 @@
+// This global variable must be set before the editor script loading.
+CKEDITOR_BASEPATH = elgg.config.wwwroot + 'mod/ckeditor/vendors/ckeditor/';
+
+require(['elgg/ckeditor', 'jquery', 'jquery.ckeditor'], function(elggCKEditor, $) {
+	$('.elgg-input-longtext:not([data-cke-init])')
+		.attr('data-cke-init', true)
+		.ckeditor(elggCKEditor.init, elggCKEditor.config);
+});
