+++ mod/developers/views/default/theme_sandbox/javascript/system_messages.php
@@ -17,17 +17,7 @@ $error = elgg_view('output/url', array(
 ));
 
 ?>
-<script type="text/javascript">
-	$(function() {
-		$('#developers-system-message').click(function() {
-			elgg.system_message('Elgg System Message');
-		})
-
-		$('#developers-error-message').click(function() {
-			elgg.register_error('Elgg Error Message');
-		})
-	});
-</script>
+<script>require(['elgg/theme_sandbox']);</script>
 
 <ul>
 	<li><?php echo $message; ?></li>
