+++ mod/messages/views/default/js/elgg/message.js
@@ -1,0 +1,14 @@
+define(function(require) {
+	var elgg = require('elgg');
+	var $ = require('jquery');
+	
+	elgg.register_hook_handler('init', 'system', function init() {
+		$("#messages-toggle").click(function() {
+			$('input[type=checkbox]').click();
+		});
+	});
+	
+	return {
+		init: init
+	};
+});
