--- mod/messages/views/default/messages/js.php
@@ -1,7 +1,0 @@
-
-// messages plugin toggle
-elgg.register_hook_handler('init', 'system', function() {
-	$("#messages-toggle").click(function() {
-		$('input[type=checkbox]').click();
-	});
-});
