+++ mod/uservalidationbyemail/start.php
@@ -45,7 +45,8 @@ function uservalidationbyemail_init() {
 	elgg_register_admin_menu_item('administer', 'unvalidated', 'users');
 
 	elgg_extend_view('css/admin', 'uservalidationbyemail/css');
-	elgg_extend_view('js/elgg', 'uservalidationbyemail/js');
+
+	elgg_require_js('elgg/uservalidationbyemail');
 
 	$action_path = dirname(__FILE__) . '/actions';
